+++ packages/components/divider/stories/divider.stories.tsx
@@ -16,6 +16,12 @@ export default {
       },
       options: ["horizontal", "vertical"],
     },
+    inset: {
+      control: {
+        type: "select",
+      },
+      options: ["none", "sm", "md"],
+    },
   },
   decorators: [
     (Story) => (